# PERF: nsmallest