```
from asv_bench.benchmarks.series_methods import *
self6 = NSort()
self6.setup("last")

%timeit self6.time_nsmallest('last')
10.6 ms ± 43.4 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)  # <- master
1.5 ms ± 17.2 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)  # <- PR
```